A package for computing the repeated measures correlation coefficient

Compute the repeated measures correlation, a statistical technique for determining the overall within-individual relationship among paired measures assessed on two or more occasions, first introduced by Bland and Altman (1995). Includes functions for diagnostics, p-value, effect size with confidence interval including optional bootstrapping, as well as graphing. Also includes several example datasets. For more details, see the web documentation https://lmarusich.github.io/rmcorr/index.html and the original paper: Bakdash and Marusich (2017) doi:10.3389/fpsyg.2017.00456 .

Compute the repeated measures correlation, a statistical technique for determining the overall within-individual relationship among paired measures assessed on two or more occasions, first introduced by Bland and Altman (1995). Includes functions for diagnostics, p-value, effect size with confidence interval including optional bootstrapping, as well as graphing. Also includes several example datasets. For more details, see the web documentation <https://lmarusich.github.io/rmcorr/index.html> and the original paper: Bakdash and Marusich (2017) <doi:10.3389/fpsyg.2017.00456>.

References

Bakdash, J.Z. & Marusich, L.R. (2017). Repeated Measures Correlation, Frontiers in Psychology, 8, 456, doi:10.3389/fpsyg.2017.00456

Bakdash, J.Z. & Marusich, L.R. (2019). Corrigendum: Repeated Measures Correlation, Frontiers in Psychology, 10, doi:10.3389/fpsyg.2019.01201

See also

Author

Maintainer: Laura R. Marusich lmarusich@gmail.com (ORCID)

Authors:

Jonathan Z. Bakdash jbakdash@gmail.com (ORCID)