A package for computing the repeated measures correlation coefficient

Compute the repeated measures correlation, a statistical technique for determining the overall within-individual relationship among paired measures assessed on two or more occasions, first introduced by Bland and Altman (1995). Includes functions for diagnostics, p-value, effect size with confidence interval including optional bootstrapping, as well as graphing. Also includes several example datasets. For more details, see Bakdash and Marusich (2017) and corrected equations in Bakdash and Marusich (2019).

References

Bakdash, J.Z. & Marusich, L.R. (2017). Repeated Measures Correlation, Frontiers in Psychology, 8, 256, doi: 10.3389/fpsyg.2017.00456 .

Bakdash, J.Z. & Marusich, L.R. (2019). Corrigendum: Repeated Measures Correlation, Frontiers in Psychology, doi: 10.3389/fpsyg.2019.01201 .